+++ packages/template-react/src/index.ts
@@ -1,8 +1,8 @@
-export * from "../../../src/template-package/render/TemplatePackageRenderer";
-export * from "../../../src/template-package/render/productRenderIdentity";
-export * from "../../../src/template-package/render/ScaledTemplatePackagePreview";
-export * from "../../../src/template-package/render/TemplateInspectionPreview";
-export * from "../../../src/template-package/render/TemplateInspectionViewport";
-export { collectTemplatePackageRenderWarnings } from "../../../src/template-package/render/packageRenderUtils";
+export * from "./render/TemplatePackageRenderer";
+export * from "./render/productRenderIdentity";
+export * from "./render/ScaledTemplatePackagePreview";
+export * from "./render/TemplateInspectionPreview";
+export * from "./render/TemplateInspectionViewport";
+export { collectTemplatePackageRenderWarnings } from "./render/packageRenderUtils";
 export * from "./runtime";
 export * from "./session";